+++ pyrf/gui/plot_tools.py
@@ -6,6 +6,8 @@ from pyrf.units import M
 from pyrf.numpy_util import calculate_channel_power
 LARGE_NEGATIVE_NUMBER = -900000
 
+# minimum size allowed for auto peak
+MIN_AUTO_POS_SIZE = 1000
 class triggerControl(pg.ROI):
     """
     Class to represent the trigger controls in the plot
@@ -127,6 +129,7 @@ class triggerControl(pg.ROI):
         else:
             self.currentPen = self.pen
         self.update()
+
 class Trace(object):
     """
     Class to represent a trace in the plot
@@ -164,11 +167,13 @@ class Trace(object):
         self.average_list = []
 
     def compute_channel_power(self):
+
         if self.calc_channel_power and not self.blank:
             if min(self.channel_power_range) > min(self.freq_range) and max(self.channel_power_range) < max(self.freq_range):
-                    min_bin = (np.abs(self.freq_range-min(self.channel_power_range))).argmin()
-                    max_bin = (np.abs(self.freq_range-max(self.channel_power_range))).argmin()
-                    self.channel_power = calculate_channel_power(self.data[min_bin:max_bin])
+                    if self.data is not None:
+                        min_bin = (np.abs(self.freq_range-min(self.channel_power_range))).argmin()
+                        max_bin = (np.abs(self.freq_range-max(self.channel_power_range))).argmin()
+                        self.channel_power = calculate_channel_power(self.data[min_bin:max_bin])
 
     def update_curve(self, xdata, ydata, usable_bins, sweep_segments):
         if self.store or self.blank:
@@ -241,7 +246,7 @@ class Marker(object):
     """
     Class to represent a marker on the plot
     """
-    def __init__(self,plot_area, marker_name, color):
+    def __init__(self,plot_area, marker_name, color, controller):
 
         self.name = marker_name
         self.marker_plot = pg.ScatterPlotItem()
@@ -249,7 +254,7 @@ class Marker(object):
         self.selected = False
         self.data_index = None
         self.xdata = []
-        self.ydata = 0
+        self.ydata = []
         self.trace_index = 0
         self.color = color
         self.draw_color = color
@@ -257,14 +262,29 @@ class Marker(object):
         self._plot = plot_area
         self.coursor_dragged = False
 
+        self.controller = controller
+
         cursor_pen = pg.mkPen((0,0,0,0), width = 40)
         self.cursor_line = InfiniteLine(pen = cursor_pen, pos = -100, angle = 90, movable = True)
         self.cursor_line.setHoverPen(pg.mkPen((0,0,0, 0), width = 40))
 
         def dragged():
-            self.data_index = np.abs( self.xdata-self.cursor_line.value()).argmin()
+            # determine index of click
+            index = np.abs(self.xdata-int(self.cursor_line.value())).argmin()
+
+            # calculate the region around the index to check for maximum value
+            index_region_offset = int(0.01 * len(self.ydata))
+            if int(min(self.xdata)) > int(min(self._plot.view_box.viewRange()[0])) or int(max(self.xdata)) > int(max(self._plot.view_box.viewRange()[0])) or len(self.ydata) < MIN_AUTO_POS_SIZE:
+                self.data_index = index
+            else:
+                # position of marker is the maximum of the region surrounding the area where the user clicked
+                if (index - index_region_offset) > 0:
+                    self.data_index = np.where(self.ydata == max(self.ydata[index - index_region_offset: index + index_region_offset]))[0]
+
             self.cursor_line.setPen(cursor_pen)
             self.draw_color = colors.MARKER_HOVER
+            self.controller.apply_plot_options(marker_dragged = True)
+            self.update_pos(self.xdata, self.ydata)
         self.cursor_line.sigDragged.connect(dragged)
 
         def hovering():
@@ -273,6 +293,7 @@ class Marker(object):
 
         def not_hovering():
             self.draw_color = color
+            self.update_pos(self.xdata, self.ydata)
         self.cursor_line.sigHoveringFinished.connect(not_hovering)
 
     def remove_marker(self, plot):
@@ -286,6 +307,7 @@ class Marker(object):
     def enable(self, plot):
         self.enabled = True
         self.add_marker(plot)
+        self.controller.apply_plot_options(marker_dragged = True)
 
     def disable(self, plot):
         self.enabled = False
@@ -295,6 +317,8 @@ class Marker(object):
 
     def update_pos(self, xdata, ydata):
 
+        # calculate scale offset for marker
+        scale = np.abs( max(self._plot.view_box.viewRange()[1]) - min(self._plot.view_box.viewRange()[1])) * 0.01
         self.marker_plot.clear()
         self._plot.window.removeItem(self.marker_plot)
         self._plot.window.addItem(self.marker_plot)
@@ -315,11 +339,12 @@ class Marker(object):
         self.ydata = ydata
         if not self.coursor_dragged:
             self.cursor_line.setValue(xpos)
+        brush_color = self.draw_color + (20,)
         self.marker_plot.addPoints(x = [xpos],
-                                   y = [ypos],
-                                    symbol = '+',
-                                    size = 25, pen = pg.mkPen(self.draw_color), 
-                                    brush = self.draw_color)
+                                   y = [ypos + scale],
+                                    symbol = 't',
+                                    size = 20, pen = pg.mkPen(self.draw_color),
+                                    brush = brush_color)
 
 
 class InfiniteLine(pg.InfiniteLine):